Lot of code fixing and modernization

+++ Network/mdz_net_chains/socket_chain_aes.cpp
@@ -48,12 +48,12 @@ void Socket_Chain_AES::setPhase1Key(const char *pass)
     SHA256_Final((unsigned char *)phase1Key, &sha256);
 }
 
-int Socket_Chain_AES::partialRead(void *data, const uint32_t &datalen)
+ssize_t Socket_Chain_AES::partialRead(void *data, const uint32_t &datalen)
 {
     if (!initialized)
         return Socket_StreamBase::partialRead(data,datalen);
 
-    int r = Socket_StreamBase::partialRead(data,datalen);
+    ssize_t r = Socket_StreamBase::partialRead(data,datalen);
     if (r<=0)
         return r;
 
@@ -69,7 +69,7 @@ int Socket_Chain_AES::partialRead(void *data, const uint32_t &datalen)
     return r;
 }
 
-int Socket_Chain_AES::partialWrite(const void *data, const uint32_t &datalen)
+ssize_t Socket_Chain_AES::partialWrite(const void *data, const uint32_t &datalen)
 {
     if (!initialized)
         return Socket_StreamBase::partialWrite(data,datalen);
@@ -93,7 +93,7 @@ int Socket_Chain_AES::partialWrite(const void *data, const uint32_t &datalen)
     writeParams.cryptoXOR(edata,datalen,true);
 
     // Try to transmit the encrypted data...
-    int r=Socket_StreamBase::partialWrite(edata,datalen);
+    ssize_t r=Socket_StreamBase::partialWrite(edata,datalen);
     if (r>0)
     {
         // Data transmited.. reduce it.
